Moved test data to pdfmake-tests.ts
